requirements.txt
pandas==2.0.3
python-docx==0.8.11
PyPDF2==3.0.1
openai==00.27.8
langchain==0.0.219
python-dateutil==2.8.2
PyQt5